Add a choice point formatter
This commit adds a CardFormatter class for choice points, with methods
returning appropriate values for the regions of a card. For example,
`ChoicePoint::CardFormatter#pre_heading` returns a string that will be
displayed in the card's "pre-heading" position, which includes the name
of the choice point's author. While decorator methods are content-
oriented (e.g., a ChoicePointDecorator would be expected to have
different methods from, say, an OptionDecorator), formatter methods are
layout-oriented. Any type of resource that we might want to format as
a card could have its own CardFormatter class, which would be expected
to implement methods like `#pre-heading`, `#heading`, `#content`, etc.

An instance of `ChoicePoint::CardFormatter` is created by calling
`ChoicePointDecorator#card`. Within formatters (as with decorators),
view helpers like `link_to` can be accessed as `h.link_to`.