Refactor choice point card
This commit rewrites the choice point card partial to take advantage
of the CardFormatter class added by the last commit. The card partial
is now resource-agnostic, meaning it is not written particularly for
choice points, but rather for any resource that happens to have a
corresponding formatter.

(This is probably overkill for present purposes, since we are unlikely
to use this card for anything but choice points, but I thought it was a
useful exercise.)