Replace 'shared' with 'application' in views
This commit replaces the "app/views/shared" directory with "app/views/
application". This means that calls to `render` that pass shared
partials no longer need to be written as "shared/navbar" (for example),
but can just say "navbar", since Rails will always look in the
"application" directory when the partial isn't found in the directory
that corresponds to the controller's name.