Speed up `choice_points#index` loading
This commit refactors the `ChoicePoint#vote_from?` and `ChoicePoint#
favorited_by?` methods and adds an `includes` query to
`ChoicePointsController#index` to speed up loading for the
`choice_points#index` action.

The commit also adds a `voters` association to `ChoicePoint`, so that
the people who voted on a particular choice point can be easily
retrieved with `ChoicePoint#voters`. For consistency, `Option#users`
is renamed to `Option#voters`.